requirements.txt
pycryptodome
cryptography
pywin32
sqlite3
